Refactor action bar, use object instead of class for logging [1/1]
[Problem]
The action bar code was all located in the main activity - this was not well encapsulated.
The logger class took a class; passing in "this" is less code and more efficient.

[Solution]
Create an action bar factory and a set of class that encapsulates the action bar and action bar tab generation.
Refactor the logger class to use Object so "this" can be passed in.