Change StudentManager into a singleton [1/1]
[Problem]
The StudentManager was being obtained by casting the activity class into a TutorTab reference. This creates unnecessary
coupling; classes should not need to reference the activity. The context is all that is needed.

[Solution]
Change the StudentManager class into a singleton. Only a context is needed to construct due to the DB.