Timer does not support stop/start or inc/dec time [1/1]
[Problem]
The timer does not support stopping and starting the timer time or increasing/decreasing the reset time at a student
    entry granularity.
Most of the time units are in seconds, which is not consistent with much of the java FWK, which is in MS.

[Solution]
Add support for stop/start and inc/dec of the timer.
Change time unit to be in MS instead of seconds within the application.