Refactor: add rest time, time left, is stopped to student entry [1/1]
[Problem]
The reset time, time left, and whether the timer is stopped is all connected to the student. Thus this information
should all be in the same data structure.

The start/stop button is really emulating a start/pause button.

[Solution]
Add reset time, time left, and whether the timer is stopped to the CurrentStudentEntry data structure.

Conflate the start/pause button into a single button that changes behavior (and text) on whether the timer is running or
not.