Issue-8: NumberPicker looks different on different devices [1/1]
[Problem]
NumberPicker looks different on different device. There is some style issue and the widget's height was defined using dp
rather than using wrap_content. When using wrap_content, the height was unreasonable, consuming a lot of screen real
estate.

[Solution]
The reset time changer is now a button, which, when clicked, launches a dialogue with the number pickers.