Issue-9: Time left text is changed by multiple students [1/1]
[Problem]
The views in the ListView are being recycled and the active student callbacks are updating the same TextView causing the
"time left" text in the student entry to flip-flop between different active student's time left.

[Solution]
Set the callback of the previous student in the ArrayAdapter.getView() to be null if the view's previous student is
targeted to update the view's time left TextView.